label updated

+++ src/main/java/com/bridgeit/fundooNote/noteservice/service/INoteService.java
@@ -7,7 +7,7 @@ import com.bridgeit.fundooNote.noteservice.model.Note;
 public interface INoteService {
 
 	public long addNote(Note note,String token );
-	public void deleteNode(long id);
+	public void deleteNode(int id);
 	public void updateNode(Note note,String token);
 	public List<Note> displayAllNote(String token);
 }
